requirements.txt
requests==2.31.0
PyYAML==6.0.1
Pillow>=10.0.0
Flask>=3.0.0
Werkzeug>=3.0.0
gunicorn>=21.2.0

